+++ node/commons/compatibility-layer.ts
@@ -1,6 +1,12 @@
 import { propOr } from "ramda";
+import VtexSeller from "./models/VtexSeller";
+import { IndexingType } from "../resolvers/products";
 
-export const convertBiggyProduct = (product: any, tradePolicy?: string) => {
+export const convertBiggyProduct = (
+  product: any,
+  tradePolicy?: string,
+  indexingType?: IndexingType,
+) => {
   const categories: string[] = product.categories
     ? product.categories.map((_: any, index: number) => {
         const subArray = product.categories.slice(0, index);
@@ -9,7 +15,7 @@ export const convertBiggyProduct = (product: any, tradePolicy?: string) => {
     : [];
 
   const skus: any[] = propOr([], "skus", product).map(
-    convertSKU(product, tradePolicy),
+    convertSKU(product, indexingType, tradePolicy),
   );
 
   return {
@@ -32,7 +38,36 @@ export const convertBiggyProduct = (product: any, tradePolicy?: string) => {
   };
 };
 
-const convertSKU = (product: any, tradePolicy?: string) => (sku: any) => {
+const getSellersIndexedByApi = (
+  product: any,
+  sku: any,
+  tradePolicy?: string,
+) => {
+  const selectedPolicy = tradePolicy
+    ? sku.policies.find((policy: any) => policy.id === tradePolicy)
+    : sku.policies[0];
+
+  const biggySellers = (selectedPolicy && selectedPolicy.sellers) || [];
+
+  return biggySellers.map((seller: any) => {
+    const price = seller.price || sku.price || product.price;
+    const oldPrice = seller.oldPrice || sku.oldPrice || product.oldPrice;
+    const installment = seller.installment || product.installment;
+
+    return new VtexSeller(seller.id, price, oldPrice, installment);
+  });
+};
+
+const getSellersIndexedByXML = (product: any) => {
+  const { installment, price, oldPrice } = product;
+  return [new VtexSeller("1", price, oldPrice, installment)];
+};
+
+const convertSKU = (
+  product: any,
+  indexingType?: IndexingType,
+  tradePolicy?: string,
+) => (sku: any) => {
   const image = {
     cacheId: product.product || product.id,
     imageId: product.product || product.id,
@@ -41,40 +76,10 @@ const convertSKU = (product: any, tradePolicy?: string) => (sku: any) => {
     imageText: "principal",
   };
 
-  const selectedPolicy = tradePolicy
-    ? sku.policies.find((policy: any) => policy.id === tradePolicy)
-    : sku.policies[0];
-  const biggySellers = (selectedPolicy && selectedPolicy.sellers) || [];
-
-  const sellers = biggySellers.map((seller: any) => {
-    const price = seller.price || sku.price || product.price;
-    const oldPrice = seller.oldPrice || sku.oldPrice || product.oldPrice;
-    const installment = seller.installment || product.installment;
-
-    return {
-      sellerId: seller.id,
-      sellerName: "",
-      commertialOffer: {
-        AvailableQuantity: 10000,
-        discountHighlights: [],
-        teasers: [],
-        Installments: installment
-          ? [
-              {
-                Value: installment.value,
-                InterestRate: 0,
-                TotalValuePlusInterestRate: price,
-                NumberOfInstallments: installment.count,
-                Name: "",
-              },
-            ]
-          : null,
-        Price: price,
-        ListPrice: oldPrice,
-        PriceWithoutDiscount: price,
-      },
-    };
-  });
+  const sellers =
+    indexingType === IndexingType.XML
+      ? getSellersIndexedByXML(product)
+      : getSellersIndexedByApi(product, sku, tradePolicy);
 
   return {
     sellers,