+++ react/components/SearchQuery.js
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { useQuery } from "react-apollo";
 import PropTypes from "prop-types";
-import { path, pathOr, isEmpty, reject } from "ramda";
+import { path, pathOr, reject, isEmpty } from "ramda";
 import { useRuntime } from "vtex.render-runtime";
 import { onSearchResult } from "vtex.sae-analytics";
 import BiggyClient from "../utils/biggy-client.ts";
@@ -95,7 +95,7 @@ const SearchQuery = ({
 
   searchQuery.variables = {
     withFacets: true,
-    query: reject(isEmpty, ["search", attributePath]).join("/"),
+    query: variables.query,
     map: map || "ft",
     orderBy: order,
     from: 0,